% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clusterprofiler_convenience.R
\name{dotplot_compareClusterResult}
\alias{dotplot_compareClusterResult}
\title{Dotplot for compareClusterResult object}
\usage{
dotplot_compareClusterResult(
  compareClusterResult,
  top_n = 5,
  facet_rgx = "\\\\S",
  facet_lvls = "foo",
  color = "p.adjust",
  size = "GeneRatio",
  xlab = "",
  ylab = "",
  fontsize = 12,
  title = "",
  cluster_lvls = 1:length(levels(droplevels(as_tibble(compareClusterResult)$Cluster))),
  print_clust_lvls = FALSE
)
}
\arguments{
\item{compareClusterResult}{compareClusterResult object}

\item{top_n}{numeric value for how many top gene sets to keep from each cluster}

\item{facet_rgx}{character value for regex to str_extract from the 'Cluster' column for facetting}

\item{facet_lvls}{character value for levels for facetting variable. Should correspond to 'facet_rgx'}

\item{color}{character for variable for dot color aesthetic}

\item{size}{character for variable for dot size aesthetic}

\item{xlab}{x axis title}

\item{ylab}{y axis title}

\item{fontsize}{font size}

\item{title}{plot title}

\item{cluster_lvls}{numeric vector to rearrange the order of the cluster levels. Use 'print_clust_lvls' to figure out the original order of the cluster levels.}

\item{print_clust_lvls}{Don't plot anything. Just print the cluster levels after dropping those with no results (significant results)}
}
\value{
GGplot object if print_clust_lvls set to FALSE
}
\description{
Dotplot for compareClusterResult object
Provides more control over the plot than the method from clusterProfiler
}
