% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/KLmisc.R
\docType{package}
\name{KLmisc}
\alias{KLmisc}
\title{KLmisc: Various functions for genomic analyses.}
\description{
Various functions for genomic analyses
}
\section{KLmisc functions}{

dotplot_compareClusterResult
}

